pyproject.toml
[tool.poetry]
name = "sisweb"
version = "0.1.0"
description = ""
authors = ["chamba <urgosxd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
dj-rest-auth = "^5.0.2"
django-allauth = "^0.58.2"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
install = "^1.3.5"
django-currentuser = "^0.6.1"
wheel = "^0.42.0"
mysqlclient = "^2.2.4"
celery = "^5.3.6"
schedule = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
